SA-903: Release new navigation
