+++ src/components/navigation/components/NavGroup.js
@@ -15,8 +15,7 @@ const NavGroup = (props) => {
     label,
     toggleMobileNav,
     mobile,
-    location,
-    newNav
+    location
   } = props;
 
   useEffect(() => {
@@ -46,7 +45,7 @@ const NavGroup = (props) => {
         <ChevronLeft className={styles.chevron}/>
       </a>
       <ul className={styles.NestedList}>
-        {children.map((child, key) => <NavItem newNav={newNav} {...child} mobile={mobile} location={location} key={key} toggleMobileNav={toggleMobileNav}/>)}
+        {children.map((child, key) => <NavItem {...child} mobile={mobile} location={location} key={key} toggleMobileNav={toggleMobileNav}/>)}
       </ul>
     </li>
   );
